Shifted this to a toggle functionality
