+++ .changeset/lucky-games-pump.md
@@ -1,0 +1,5 @@
+---
+"@atomicsmash/cli": major
+---
+
+Added new toggle media proxy function
